requirements.txt
python-telegram-bot
requests
telebot